requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
langchain-community
chromadb
langserve